+++ banzai/bpm.py
@@ -11,9 +11,6 @@ logger = logging.getLogger(__name__)
 
 
 class BPMUpdater(Stage):
-    @property
-    def group_by_attributes(self):
-        return None
 
     def do_stage(self, images):
         images_to_remove = []