+++ banzai/gain.py
@@ -9,10 +9,6 @@ class GainNormalizer(Stage):
     def __init__(self, pipeline_context):
         super(GainNormalizer, self).__init__(pipeline_context)
 
-    @property
-    def group_by_attributes(self):
-        return None
-
     def do_stage(self, images):
         images_to_remove = []
         for image in images: