+++ banzai/tests/test_dark_comparer.py
@@ -23,9 +23,9 @@ def test_no_input_images():
     assert len(images) == 0
 
 
-def test_group_by_keywords():
+def test_master_selection_criteria():
     comparer = DarkComparer(None)
-    assert comparer.group_by_attributes == ['ccdsum']
+    assert comparer.master_selection_criteria == ['ccdsum']
 
 
 @mock.patch('banzai.calibrations.Image')