+++ banzai/tests/test_flat_comparer.py
@@ -23,9 +23,9 @@ def test_no_input_images(set_random_seed):
     assert len(images) == 0
 
 
-def test_group_by_keywords(set_random_seed):
+def test_master_selection_criteria(set_random_seed):
     comparer = FlatComparer(FakeContext())
-    assert comparer.group_by_attributes == ['ccdsum', 'filter']
+    assert comparer.master_selection_criteria == ['ccdsum', 'filter']
 
 
 @mock.patch('banzai.calibrations.Image')